Pass test to write and read simple.

+++ src/flume_log.rs
@@ -9,9 +9,9 @@ pub struct StreamOpts{
 pub trait FlumeLog 
 {
     //TODO: errors.
-    fn get(&mut self, seq_num: usize) -> Result<Vec<u8>, ()>;
-    fn clear(&mut self, seq_num: usize);
-    fn latest(&self) -> usize;
-    fn append(& mut self, buff: &[u8]) -> Result<usize, ()>;
+    fn get(&mut self, seq_num: u64) -> Result<Vec<u8>, ()>;
+    fn clear(&mut self, seq_num: u64);
+    fn latest(&self) -> u64;
+    fn append(& mut self, buff: &[u8]) -> Result<u64, ()>;
 }
