+++ src/mem_log.rs
@@ -18,17 +18,17 @@ impl MemLog {
 
 impl FlumeLog for MemLog {
     //TODO: errors.
-    fn get(&mut self, seq_num: usize) -> Result<Vec<u8>, ()> {
-        let slice = self.log.get(seq_num).ok_or(())?;
+    fn get(&mut self, seq_num: u64) -> Result<Vec<u8>, ()> {
+        let slice = self.log.get(seq_num as usize).ok_or(())?;
         Ok(slice.clone())
     }
-    fn clear(&mut self, seq: usize) {
-        self.log[seq] = Vec::new();
+    fn clear(&mut self, seq: u64) {
+        self.log[seq as usize] = Vec::new();
     }
-    fn latest(&self) -> usize {
-        self.log.len()
+    fn latest(&self) -> u64 {
+        self.log.len() as u64
     }
-    fn append(& mut self, buff: &[u8]) -> Result<usize, ()> {
+    fn append(& mut self, buff: &[u8]) -> Result<u64, ()> {
 
         let seq = self.latest();
         let mut vec = Vec::new();
